NetService removal refactoring
Moved NetService GetSortedLanes into ExtSegmentManager.
Moved LanePos into TrafficManager.Manager.Impl namespace.
