+++ TLM/TLM/Manager/Impl/ExtSegmentManager.cs
@@ -1,6 +1,8 @@
 namespace TrafficManager.Manager.Impl {
     using ColossalFramework;
     using CSUtil.Commons;
+    using System;
+    using System.Collections.Generic;
     using TrafficManager.API.Manager;
     using TrafficManager.API.Traffic.Data;
     using TrafficManager.State.ConfigData;
@@ -259,6 +261,117 @@ namespace TrafficManager.Manager.Impl {
             }
 
             return new GetSegmentLaneIdsEnumerable(initialLaneId, netInfo.m_lanes.Length, laneBuffer);
+        }
+
+        /// <summary>
+        /// Assembles a geometrically sorted list of lanes for the given segment.
+        /// If the <paramref name="startNode"/> parameter is set only lanes supporting traffic to flow towards the given node are added to the list, otherwise all matched lanes are added.
+        /// </summary>
+        /// <param name="segmentId">segment id</param>
+        /// <param name="segment">segment data</param>
+        /// <param name="startNode">reference node (optional)</param>
+        /// <param name="laneTypeFilter">lane type filter, lanes must match this filter mask</param>
+        /// <param name="vehicleTypeFilter">vehicle type filter, lanes must match this filter mask</param>
+        /// <param name="reverse">if true, lanes are ordered from right to left (relative to the
+        ///     segment's start node / the given node), otherwise from left to right</param>
+        /// <param name="sort">if false, no sorting takes place
+        ///     regardless of <paramref name="reverse"/></param>
+        /// <returns>sorted list of lanes for the given segment</returns>
+        public IList<LanePos> GetSortedLanes(ushort segmentId,
+                                             ref NetSegment segment,
+                                             bool? startNode,
+                                             NetInfo.LaneType? laneTypeFilter = null,
+                                             VehicleInfo.VehicleType? vehicleTypeFilter = null,
+                                             bool reverse = false,
+                                             bool sort = true) {
+            // TODO refactor together with getSegmentNumVehicleLanes, especially the vehicle type and lane type checks
+            NetManager netManager = Singleton<NetManager>.instance;
+            var laneList = new List<LanePos>();
+
+            bool inverted = (segment.m_flags & NetSegment.Flags.Invert) != NetSegment.Flags.None;
+
+            NetInfo.Direction? filterDir = null;
+            NetInfo.Direction sortDir = NetInfo.Direction.Forward;
+
+            if (startNode != null) {
+                filterDir = (bool)startNode
+                                ? NetInfo.Direction.Backward
+                                : NetInfo.Direction.Forward;
+                filterDir = inverted
+                                ? NetInfo.InvertDirection((NetInfo.Direction)filterDir)
+                                : filterDir;
+                sortDir = NetInfo.InvertDirection((NetInfo.Direction)filterDir);
+            } else if (inverted) {
+                sortDir = NetInfo.Direction.Backward;
+            }
+
+            if (reverse) {
+                sortDir = NetInfo.InvertDirection(sortDir);
+            }
+
+            NetInfo segmentInfo = segment.Info;
+            uint curLaneId = segment.m_lanes;
+            byte laneIndex = 0;
+
+            while (laneIndex < segmentInfo.m_lanes.Length && curLaneId != 0u) {
+                NetInfo.Lane laneInfo = segmentInfo.m_lanes[laneIndex];
+                if ((laneTypeFilter == null ||
+                     (laneInfo.m_laneType & laneTypeFilter) != NetInfo.LaneType.None) &&
+                    (vehicleTypeFilter == null || (laneInfo.m_vehicleType & vehicleTypeFilter) !=
+                     VehicleInfo.VehicleType.None) &&
+                    (filterDir == null ||
+                     segmentInfo.m_lanes[laneIndex].m_finalDirection == filterDir)) {
+                    laneList.Add(
+                        new LanePos(
+                            curLaneId,
+                            laneIndex,
+                            segmentInfo.m_lanes[laneIndex].m_position,
+                            laneInfo.m_vehicleType,
+                            laneInfo.m_laneType));
+                }
+
+                curLaneId = netManager.m_lanes.m_buffer[curLaneId].m_nextLane;
+                ++laneIndex;
+            }
+
+            if (sort) {
+                int CompareLanePositionsFun(LanePos x, LanePos y) {
+                    bool fwd = sortDir == NetInfo.Direction.Forward;
+                    if (Math.Abs(x.position - y.position) < 1e-12) {
+                        if (x.position > 0) {
+                            // mirror type-bound lanes (e.g. for coherent disply of lane-wise speed limits)
+                            fwd = !fwd;
+                        }
+
+                        if (x.laneType == y.laneType) {
+                            if (x.vehicleType == y.vehicleType) {
+                                return 0;
+                            }
+
+                            if ((x.vehicleType < y.vehicleType) == fwd) {
+                                return -1;
+                            }
+
+                            return 1;
+                        }
+
+                        if ((x.laneType < y.laneType) == fwd) {
+                            return -1;
+                        }
+
+                        return 1;
+                    }
+
+                    if (x.position < y.position == fwd) {
+                        return -1;
+                    }
+
+                    return 1;
+                }
+
+                laneList.Sort(CompareLanePositionsFun);
+            }
+            return laneList;
         }
 
         protected override void InternalPrintDebugInfo() {
