+++ TLM/TLM/TLM.csproj
@@ -150,6 +150,7 @@
     <Compile Include="Manager\Impl\GetSegmentLaneIdsEnumerable.cs" />
     <Compile Include="Manager\Impl\GetSegmentLaneIdsEnumerator.cs" />
     <Compile Include="Manager\Impl\LaneIdAndIndex.cs" />
+    <Compile Include="Manager\Impl\LanePos.cs" />
     <Compile Include="Notifier.cs" />
     <Compile Include="Patch\CustomPathFindPatchAttribute.cs" />
     <Compile Include="Patch\PatchCommons.cs" />
