+++ TLM/TLM/Util/Record/SegmentRecord.cs
@@ -74,7 +74,8 @@ namespace TrafficManager.Util.Record {
         }
 
         public static List<uint> GetAllLanes(ushort segmentId) {
-            var lanes =  NetService.Instance.GetSortedLanes(
+            ExtSegmentManager extSegmentManager = ExtSegmentManager.Instance;
+            var lanes = extSegmentManager.GetSortedLanes(
                 segmentId,
                 ref segmentId.ToSegment(),
                 startNode: null,
